requirements.txt
pandas==2.1.1
numpy==1.24.3
python-binance==1.0.19
websocket-client==1.6.3
plotly==5.17.0
python-dotenv==1.0.0
ta==0.10.2